Fleshed out data structure and code to check whether it is a latin-square.
Hand-made some latin squares by hand to check that code; have not put
in searching code to find latin squares in general.

Also, the current instance of the data structure has only been shown
to represent a full (potentially non-latin) square, which you operate
on by swapping elements.  Later I might want to put in a more general
instance that allows for unfilled slots.  (Hopefully that can be done
as a more general instance of what is already here.)