Added separated-variant formulation of Latin-Squares.
The essential idea is to apply the same "no repeats in row or columns"
on the left and right squares, and then add the additional constraint
that the left and right squares are "orthogonal", which means that no
pair appears twice in their product `left * right`.

Addressing this made me realize that the previous `is_latin` had a
bug, in that it assumed that the input would eliminate duplicates by
construction, which of course is not generally enforced and thus
should be checked.

Added additional way to print the product squares, which hold tuples,
but without using the more verbose `(x, y)` output format.  Achieved
this by adding a level-of-indirection via a `ElemToStr` trait.