Added the other column selection policy noted by Knuth.
Also added instrumentation to gather update_count statistics.

However, my current numbers do not agree with Knuth's.
With the "branch-minimizing" strategy, I count 37 updates (vs Knuth's 28).
With the S-heuristic ignoring strategy, I count 42 updates (vs Knuth's 33).

I guess the fact that both numbers differ by five is a hint of something.

It is not clear from the paper whether Knuth considers the unlinking
of a column header to be one of the kinds of updates of interest here.
If I remove column updates from the count, my numbers change to 18 and
20 for the two respective cases.

it seems like I must be missing something.